+++ src/app/core/models/entity/lear-credential.ts
@@ -73,7 +73,7 @@ export interface Power {
   type: string;
 }
 
-
+//todo remove "Tmf" prefix
 export type TmfFunction = 'Onboarding' | 'ProductOffering' | 'Certification' | 'CredentialIssuer' | 'Login';
 export type TmfAction = 'Execute' | 'Create' | 'Update' | 'Delete' | 'Upload' | 'Attest' | 'Configure' | 'oidc_m2m'