+++ src/app/features/credential-details/components/details-power/details-power.component.ts
@@ -3,6 +3,7 @@ import { CapitalizePipe } from './../../../../shared/pipes/capitalize.pipe';
 import { Component, inject, InjectionToken } from '@angular/core';
 import { FunctionActions } from '../../helpers/credential-details-helpers';
 import { TranslatePipe } from '@ngx-translate/core';
+import { environment } from 'src/environments/environment';
 
 export const detailsPowerToken = new InjectionToken<FunctionActions[]>('DETAILS_POWER');
 
@@ -15,5 +16,7 @@ export const detailsPowerToken = new InjectionToken<FunctionActions[]>('DETAILS_
 })
 export class DetailsPowerComponent {
   public powers: FunctionActions[] = inject(detailsPowerToken);
+  // if at some point more than one domain is possible, it should be extracted from each power
+  public domain: string = environment.sys_tenant;
 
 }