+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.3](https://github.com/in2workspace/in2-issuer-ui/releases/tag/v2.1.3)
+### Added
+- Added environment variable `sys_admin` to set credential powers "domain" field and display it in issuance form and credential details page.
+
+### Fixed
+- Added translations for the country selector.
+
 ## [2.1.2](https://github.com/in2workspace/in2-issuer-ui/releases/tag/v2.1.2)
 ### Added
 - Admin organization identifier is now configurable.